requirements.txt
discord.py==2.6.2
openai==1.102.0
psutil==7.0.0
python-dotenv==1.1.1

tiktoken==0.5.2
Pillow==10.3.0
aiofiles==23.2.1
aiohttp==3.9.5
